Voice-Assistant/Voice-Assistant-master copy/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "groq>=0.29.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sarvamai>=0.1.10",
]
